# packages/contracts/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LittleArc API",
    "version": "0.1.0",
    "description": "Versioned LittleArc API contract through OFF-04 synchronization."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local synthetic development API"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "consumerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
...
  },
  "paths": {
    "/v1/sync": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Pull ordered household changes or request reconciliation",
        "security": [
          {
            "consumerSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered changes or a typed reset requirement",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "changes": {
                          "type": "array",
                          "items": {
                            "oneOf": [
                              {
                                "type": "object",
                                "properties": {
                                  "changedAt": {
                                    "type": "string",
                                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$(?![\\s\\S])",
                                    "description": "Normalized UTC ISO 8601 timestamp"
                                  },
                                  "entity": {
                                    "type": "object",
                                    "properties": {
                                      "childId": {
                                        "type": "string",
                                        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$(?![\\s\\S])",
                                        "description": "UUIDv7 identifier"
                                      },
                                      "dateOfBirth": {
                                        "type": "string",
                                        "format": "date"
                                      },
                                      "preferredName": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 120
                                      },
                                      "revision": {
                                        "type": "integer",
                                        "exclusiveMinimum": 0,
                                        "description": "Mutable entity revision"
                                      },
                                      "updatedAt": {
                                        "type": "string",
                                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$(?![\\s\\S])",
                                        "description": "Normalized UTC ISO 8601 timestamp"
                                      }
                                    },
                                    "required": [
                                      "childId",
                                      "dateOfBirth",
                                      "preferredName",
                                      "revision",
                                      "updatedAt"
                                    ]
                                  },
                                  "entityId": {
                                    "type": "string",
                                    "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$(?![\\s\\S])",
                                    "description": "UUIDv7 identifier"
                                  },
                                  "entityType": {
                                    "type": "string",
                                    "enum": [
                                      "child"
                                    ]
                                  },
                                  "operation": {
                                    "type": "string",
                                    "enum": [
                                      "upsert"
                                    ]
                                  },
                                  "revision": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0,
                                    "description": "Mutable entity revision"
                                  },
                                  "sequence": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  }
                                },
                                "required": [
                                  "changedAt",
                                  "entity",
                                  "entityId",
                                  "entityType",
                                  "operation",
                                  "revision",
                                  "sequence"
                                ]
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "changedAt": {
                                    "type": "string",
                                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$(?![\\s\\S])",
                                    "description": "Normalized UTC ISO 8601 timestamp"
                                  },
                                  "entityId": {
                                    "type": "string",
                                    "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$(?![\\s\\S])",
                                    "description": "UUIDv7 identifier"
                                  },
                                  "entityType": {
                                    "type": "string",
                                    "enum": [
                                      "child"
                                    ]
                                  },
                                  "operation": {
                                    "type": "string",
                                    "enum": [
                                      "delete"
                                    ]
                                  },
                                  "revision": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0,
                                    "description": "Mutable entity revision"
                                  },
                                  "sequence": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  }
                                },
                                "required": [
                                  "changedAt",
                                  "entityId",
                                  "entityType",
                                  "operation",
                                  "revision",
                                  "sequence"
                                ]
                              }
                            ]
                          }
                        },
                        "hasMore": {
                          "type": "boolean"
                        },
                        "kind": {
                          "type": "string",
                          "enum": [
                            "changes"
                          ]
                        },
                        "nextCursor": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_-]{12,512}$(?![\\s\\S])",
                          "description": "Opaque pagination or synchronization cursor"
                        },
                        "serverTime": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$(?![\\s\\S])",
                          "description": "Normalized UTC ISO 8601 timestamp"
                        }
                      },
                      "required": [
                        "changes",
                        "hasMore",
                        "kind",
                        "nextCursor",
                        "serverTime"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "kind": {
                          "type": "string",
                          "enum": [
                            "resetRequired"
                          ]
                        },
                        "reason": {
                          "type": "string",
                          "enum": [
                            "initialSync",
                            "cursorExpired"
                          ]
                        },
                        "serverTime": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$(?![\\s\\S])",
                          "description": "Normalized UTC ISO 8601 timestamp"
                        }
                      },
                      "required": [
                        "kind",
                        "reason",
                        "serverTime"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Cursor or page request is invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "bad_request",
                        "contract_not_found",
                        "forbidden",
                        "not_found",
                        "record_revision_conflict",
                        "request_validation_failed",
                        "unauthorized",
                        "unknown_error"
                      ],
                      "description": "Stable application error code"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "description": "Request correlation ID"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      },
                      "default": []
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "description": "RFC Problem Details with LittleArc error code and request ID"
                }
              }
            }
          },
          "401": {
            "description": "Consumer session required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "bad_request",
                        "contract_not_found",
                        "forbidden",
                        "not_found",
                        "record_revision_conflict",
                        "request_validation_failed",
                        "unauthorized",
                        "unknown_error"
                      ],
                      "description": "Stable application error code"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "description": "Request correlation ID"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      },
                      "default": []
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "description": "RFC Problem Details with LittleArc error code and request ID"
                }
              }
            }
          },
          "403": {
            "description": "Active household membership required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "bad_request",
                        "contract_not_found",
                        "forbidden",
                        "not_found",
                        "record_revision_conflict",
                        "request_validation_failed",
                        "unauthorized",
                        "unknown_error"
                      ],
                      "description": "Stable application error code"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "description": "Request correlation ID"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      },
                      "default": []
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "description": "RFC Problem Details with LittleArc error code and request ID"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}